Don't clean up an uninitialized MprWaitHandler
It is possible for an MprWaitHandler to be created, but not added to the global
WaitService (for example, if the maximum connection count of FD_SETSIZE is
reached). In this situation, the wait handler's service field will be NULL, and
mprRemoveWaitHandler() will crash when it tries to clean up during garbage
collection.

This patch returns early from mprRemoveWaitHandler() if the current wait
handler has not been fully initialized.